Add forum install functional test
Scenario: The user should be notified on all user accessible pages, that
the installation is not completed yet.
Scenario: The user should be able to access the 'about' message when he
enters the installer.
Scenario: The user should be able to change the language of the
installer.
Scenario: The installer should not proceed if the user doesn't accept
the license agreements.
Scenario: The installer should not proceed if the user does enter
invalid database credentials.
Scenario: The installer should not proceed if the user enters an invalid
database host or if the database host is not reachable.
Scenario: The installer should not proceed if the user enters an invalid
database name or if the user can't create a database.
Scenario: The installer should not proceed if the user enters an invalid
table prefix.
Scenario: The installer should not proceed if the user sets an empty
administrator name.
Scenario: Successfully finish the installation.
Scenario: The installer should not proceed if the forum is already
installed.