Add user authentication functional test
Scenario: Login with an empty username.
Scenario: Login with an empty password.
Scenario: Login with invalid credentials.
Scenario: Flood the login form.
Scenario: Login shortly before the flood protection times out.
Scenario: Login shortly after the flood protection times out.
Scenario: Login as default group user.
Scenario: Login as a not yet activated user.
Scenario: Login as an administrator.
Scenario: Login on the admin panel without regular login and as a
regular user.
Scenario: Login on the admin panel with regular login and as a
regular user.
Scenario: Login on the admin panel without regular login and as an
administrator.
Scenario: Login on the admin panel with regular login and as an
administrator.