Add user create functional test
Scenario: Create a user with an empty name.
Scenario: Create a duplicate user name.
Scenario: Create a duplicate user email.
Scenario: Create a user with an invalid email host.
Scenario: Create a user with a weak password.
Scenario: Create a user with a mismatching password.
Scenario: Successfully create an user.